+++ docs/release-notes.md
@@ -28,6 +28,7 @@ nav_order: 9
 - Fix the Butane root partition constraint check to only examine subsequent partitions ([#2304](https://github.com/coreos/ignition/pull/2304))
 - Close `/proc/mounts` after checking whether block devices are mounted, preventing file descriptor leaks during disk setup ([#2289](https://github.com/coreos/ignition/pull/2289))
 - Drop supplementary groups when dropping privileges to write files as a user ([#2242](https://github.com/coreos/ignition/issues/2242))
+- butane: report the friendly input name (`<stdin>` instead of `/dev/stdin`) in stdin read errors
 
 ## Ignition 2.27.0 (2026-08-26)